+++ http4k-client/helidon/src/main/kotlin/org/http4k/client/HelidonClient.kt
@@ -1,10 +1,9 @@
 package org.http4k.client
 
-import io.helidon.common.http.Http
-import io.helidon.nima.webclient.ClientResponse
-import io.helidon.nima.webclient.RuntimeUnknownHostException
-import io.helidon.nima.webclient.WebClient
-import io.helidon.nima.webclient.http1.Http1Client
+import io.helidon.http.HeaderNames
+import io.helidon.http.Method
+import io.helidon.webclient.api.HttpClientResponse
+import io.helidon.webclient.api.WebClient
 import org.http4k.core.BodyMode
 import org.http4k.core.BodyMode.Memory
 import org.http4k.core.HttpHandler
@@ -16,35 +15,34 @@ import org.http4k.core.Status.Companion.UNKNOWN_HOST
 import java.io.UncheckedIOException
 import java.net.ConnectException
 import java.net.SocketTimeoutException
+import java.net.UnknownHostException
 
 object HelidonClient {
     @JvmStatic
     @JvmOverloads
     @JvmName("create")
     operator fun invoke(
-        client: Http1Client = WebClient.builder().followRedirect(false).build(),
+        client: WebClient = WebClient.builder().followRedirects(false).build(),
         bodyMode: BodyMode = Memory,
     ): HttpHandler = object : HttpHandler {
-
         override fun invoke(request: Request) = try {
             client.makeHelidonRequest(request).asHttp4k()
         } catch (e: UncheckedIOException) {
             when (e.cause) {
+                is UnknownHostException -> Response(UNKNOWN_HOST.toClientStatus(e))
                 is ConnectException -> Response(UNKNOWN_HOST.toClientStatus(e))
                 is SocketTimeoutException -> Response(CLIENT_TIMEOUT.toClientStatus(e))
                 else -> throw e
             }
-        } catch (e: RuntimeUnknownHostException) {
-            Response(UNKNOWN_HOST.toClientStatus(e))
         }
 
-        private fun Http1Client.makeHelidonRequest(request: Request) = request.headers.groupBy { it.first }
+        private fun WebClient.makeHelidonRequest(request: Request) = request.headers.groupBy { it.first }
             .entries
-            .fold(method(Http.Method.create(request.method.name)).uri(request.uri.toString())) { acc, next ->
-                acc.header(Http.Header.create(Http.Header.createName(next.key, next.key), next.value.map { it.second }))
+            .fold(method(Method.create(request.method.name)).uri(request.uri.toString())) { acc, next ->
+                acc.header(HeaderNames.create(next.key, next.key), next.value.map { it.second })
             }.submit(request.body.payload.array())
 
-        private fun ClientResponse.asHttp4k() =
+        private fun HttpClientResponse.asHttp4k() =
             headers().fold(Response(Status(status().code(), status().reasonPhrase()))) { acc, next ->
                 next.allValues().fold(acc) { acc2, value -> acc2.header(next.name(), value) }
             }.body(bodyMode(inputStream()))