+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
+### v5.9.0.0
+- **http4k-*** : Upgrade some dependency versions
+- **http4k-server-helidon** : [Breaking[ Upgrade to stable v4 of Helidon, API changes
+- **http4k-client-helidon** : [Breaking[ Upgrade to stable v4 of Helidon, API changes
+
 ### v5.8.6.0
 - **http4k-*** : Upgrade some dependency versions.
 - **http4k-core** : New filter to validate zipkin headers. H/T @time4tea